Rename MDR classifier config file
